reorg of components to dirs; renamed Search to SearchContainer for better identification of responsibility; re-applied API call in search action will need to add an API client for proper authorization
