+++ src/app/components/SearchContainer/SearchContainer.jsx
@@ -2,15 +2,15 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import SearchForm from './SearchForm';
-import SearchResultsList from './SearchResultsList';
+import SearchForm from '../SearchForm/SearchForm';
+import SearchResults from '../SearchResults/SearchResults';
 import * as searchActions from '../../actions/SearchActions';
 
-class Search extends React.Component {
+class SearchContainer extends React.Component {
   constructor(props) {
     super(props);
     const { dispatch } = props;
-    this.state = props;
+
     this.boundActions = bindActionCreators(searchActions, dispatch);
   }
 
@@ -22,25 +22,25 @@ class Search extends React.Component {
           searchField={this.props.searchField}
           {...this.boundActions}
         />
-        <SearchResultsList results={this.props.searchResults} />
+        <SearchResults results={this.props.searchResults} />
       </div>
     );
   }
 }
 
-Search.propTypes = {
+SearchContainer.propTypes = {
   searchResults: PropTypes.object,
   searchQuery: PropTypes.string,
   searchField: PropTypes.string,
 };
 
-Search.defaultProps = {
+SearchContainer.defaultProps = {
   searchResults: {},
   searchQuery: '',
   searchField: 'q',
 };
 
-Search.contextTypes = {
+SearchContainer.contextTypes = {
   router: PropTypes.object,
 };
 
@@ -55,4 +55,4 @@ const mapStateToProps = (state, ownProps) => {
 export default connect(
   mapStateToProps,
   null,
-)(Search);
+)(SearchContainer);
