+++ src/app/components/SearchResults/ResultsRow.jsx
